Added Explosive Collectable
Added collectable that damages the player when touched.
