+++ Assets/Animations/ExplosiveCollectable.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3022cd9b1bd856d42ad47ca947426012
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
