+++ Assets/Animations/ExplosiveCollectable_anim.anim
@@ -1,0 +1,98 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!74 &7400000
+AnimationClip:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: ExplosiveCollectable_anim
+  serializedVersion: 6
+  m_Legacy: 0
+  m_Compressed: 0
+  m_UseHighQualityCurve: 1
+  m_RotationCurves: []
+  m_CompressedRotationCurves: []
+  m_EulerCurves: []
+  m_PositionCurves: []
+  m_ScaleCurves: []
+  m_FloatCurves: []
+  m_PPtrCurves:
+  - curve:
+    - time: 0
+      value: {fileID: -756354950, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - time: 0.1
+      value: {fileID: -1663983557, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - time: 0.2
+      value: {fileID: 411176780, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - time: 0.3
+      value: {fileID: 826496978, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - time: 0.4
+      value: {fileID: -914754715, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - time: 0.5
+      value: {fileID: -17354639, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - time: 0.6
+      value: {fileID: -914754715, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - time: 0.7
+      value: {fileID: 826496978, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - time: 0.8
+      value: {fileID: 826496978, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - time: 0.9
+      value: {fileID: 411176780, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - time: 1
+      value: {fileID: -1663983557, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    attribute: m_Sprite
+    path: 
+    classID: 212
+    script: {fileID: 0}
+  m_SampleRate: 60
+  m_WrapMode: 0
+  m_Bounds:
+    m_Center: {x: 0, y: 0, z: 0}
+    m_Extent: {x: 0, y: 0, z: 0}
+  m_ClipBindingConstant:
+    genericBindings:
+    - serializedVersion: 2
+      path: 0
+      attribute: 0
+      script: {fileID: 0}
+      typeID: 212
+      customType: 23
+      isPPtrCurve: 1
+    pptrCurveMapping:
+    - {fileID: -756354950, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - {fileID: -1663983557, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - {fileID: 411176780, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - {fileID: 826496978, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - {fileID: -914754715, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - {fileID: -17354639, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - {fileID: -914754715, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - {fileID: 826496978, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - {fileID: 826496978, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - {fileID: 411176780, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+    - {fileID: -1663983557, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+  m_AnimationClipSettings:
+    serializedVersion: 2
+    m_AdditiveReferencePoseClip: {fileID: 0}
+    m_AdditiveReferencePoseTime: 0
+    m_StartTime: 0
+    m_StopTime: 1.0166667
+    m_OrientationOffsetY: 0
+    m_Level: 0
+    m_CycleOffset: 0
+    m_HasAdditiveReferencePose: 0
+    m_LoopTime: 1
+    m_LoopBlend: 0
+    m_LoopBlendOrientation: 0
+    m_LoopBlendPositionY: 0
+    m_LoopBlendPositionXZ: 0
+    m_KeepOriginalOrientation: 0
+    m_KeepOriginalPositionY: 1
+    m_KeepOriginalPositionXZ: 0
+    m_HeightFromFeet: 0
+    m_Mirror: 0
+  m_EditorCurves: []
+  m_EulerEditorCurves: []
+  m_HasGenericRootTransform: 0
+  m_HasMotionFloatCurves: 0
+  m_Events: []
