+++ Assets/Animations/ExplosiveCollectable_anim.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ee5d72ab9be91394b8154ec219059639
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
