+++ Assets/Prefabs/Powerup/ExplosiveCollectable.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &5503697896774214530
+--- !u!1 &7676970943839480874
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,40 +8,40 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 5503697896774214591}
-  - component: {fileID: 5503697896774214590}
-  - component: {fileID: 5503697896774214529}
-  - component: {fileID: 5503697896774214528}
-  - component: {fileID: 5503697896774214531}
+  - component: {fileID: 7676970943839480855}
+  - component: {fileID: 7676970943839480854}
+  - component: {fileID: 7676970943839480873}
+  - component: {fileID: 7676970943839480872}
+  - component: {fileID: 7676970943839480878}
   m_Layer: 0
-  m_Name: HealthPowerup
+  m_Name: ExplosiveCollectable
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &5503697896774214591
+--- !u!4 &7676970943839480855
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5503697896774214530}
+  m_GameObject: {fileID: 7676970943839480874}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 0.5, y: 0.5, z: 0.5}
+  m_LocalPosition: {x: -3.74, y: 6.04, z: 0}
+  m_LocalScale: {x: 0.35, y: 0.35, z: 0.35}
   m_ConstrainProportionsScale: 1
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!212 &5503697896774214590
+--- !u!212 &7676970943839480854
 SpriteRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5503697896774214530}
+  m_GameObject: {fileID: 7676970943839480874}
   m_Enabled: 1
   m_CastShadows: 0
   m_ReceiveShadows: 0
@@ -76,8 +76,8 @@ SpriteRenderer:
   m_SortingLayerID: -1091817389
   m_SortingLayer: 1
   m_SortingOrder: 0
-  m_Sprite: {fileID: 7482667652216324306, guid: 311925a002f4447b3a28927169b83ea6, type: 3}
-  m_Color: {r: 1, g: 0, b: 0, a: 1}
+  m_Sprite: {fileID: -756354950, guid: b04d4b1241626894bab90c2549d42867, type: 3}
+  m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
@@ -87,13 +87,13 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!61 &5503697896774214529
+--- !u!61 &7676970943839480873
 BoxCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5503697896774214530}
+  m_GameObject: {fileID: 7676970943839480874}
   m_Enabled: 1
   m_Density: 1
   m_Material: {fileID: 0}
@@ -104,7 +104,7 @@ BoxCollider2D:
   m_SpriteTilingProperty:
     border: {x: 0, y: 0, z: 0, w: 0}
     pivot: {x: 0.5, y: 0.5}
-    oldSize: {x: 1, y: 1}
+    oldSize: {x: 5.004888, y: 5}
     newSize: {x: 5.12, y: 5.12}
     adaptiveTilingThreshold: 0.5
     drawMode: 0
@@ -113,17 +113,17 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 3.2301922, y: 3.277505}
   m_EdgeRadius: 0
---- !u!95 &5503697896774214528
+--- !u!95 &7676970943839480872
 Animator:
   serializedVersion: 4
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5503697896774214530}
-  m_Enabled: 0
+  m_GameObject: {fileID: 7676970943839480874}
+  m_Enabled: 1
   m_Avatar: {fileID: 0}
-  m_Controller: {fileID: 9100000, guid: a7d52a0e2323f9d40b7a1df38d2b6229, type: 2}
+  m_Controller: {fileID: 9100000, guid: 3022cd9b1bd856d42ad47ca947426012, type: 2}
   m_CullingMode: 0
   m_UpdateMode: 0
   m_ApplyRootMotion: 0
@@ -133,19 +133,21 @@ Animator:
   m_HasTransformHierarchy: 1
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorControllerStateOnDisable: 0
---- !u!114 &5503697896774214531
+--- !u!114 &7676970943839480878
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5503697896774214530}
+  m_GameObject: {fileID: 7676970943839480874}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 7f4f5d9970c7f444a9925167e37f8b95, type: 3}
+  m_Script: {fileID: 11500000, guid: 3946060d371a6c047a397f2ff8266894, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   _speed: 3
   _bounds_Y: -10
-  _collectableID: 4
-  _audioClip: {fileID: 8300000, guid: 5ee343246919d2b4bba63ae629e95494, type: 3}
+  _collectableID: 6
+  _audioClip: {fileID: 8300000, guid: a885c043fab7e46438bba03ba9dab6bc, type: 3}
+  _rotateSpeed: 1
+  _explosionPrefab: {fileID: 6454016076013425788, guid: 2c585908bd0f3914dbaeb1701918f502, type: 3}
