+++ Assets/Prefabs/Powerup/ExplosiveCollectable.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 41217354ef7c81c4cae76c6867a36721
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
