+++ Assets/Prefabs/Powerup/ShieldPowerup.prefab
@@ -147,5 +147,5 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _speed: 3
   _bounds_Y: -10
-  _powerupID: 2
+  _collectableID: 2
   _audioClip: {fileID: 8300000, guid: 5ee343246919d2b4bba63ae629e95494, type: 3}
