+++ Assets/Scripts/Explosive.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3946060d371a6c047a397f2ff8266894
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
