+++ Assets/Sprites/Power_Ups/Explosive.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c667a500d064e3f40a8116683538a704
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
